+++ cranberry-commons/src/main/java/io/github/ololx/cranberry/commons/annotation/IncludeVarsLocal.java
@@ -1,9 +1,6 @@
 package io.github.ololx.cranberry.commons.annotation;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * CMarks a class or method so that the cranberry can run the annotation processor when only local variables are annotated.
@@ -12,11 +9,13 @@ import java.lang.annotation.Target;
  * but cannot process them if no klyukovka annotations were specified for non-local variables.
  *
  */
+@Deprecated
 @Retention(RetentionPolicy.SOURCE)
 @Target({
         ElementType.METHOD,
         ElementType.TYPE,
-        ElementType.ANNOTATION_TYPE
+        ElementType.ANNOTATION_TYPE,
+        ElementType.TYPE_USE
 })
 public @interface IncludeVarsLocal {
 }